account_service/account/serializers.py
from rest_framework.serializers import ModelSerializer
from .models import Account


class AccountSerializer(ModelSerializer):
    class Meta:
        model = Account
        fields = ['account_number', 'balance', 'owner_id']
